+++ src/main/java/org/jabref/gui/entryeditor/PreviewTab.java
@@ -1,13 +1,14 @@
 package org.jabref.gui.entryeditor;
 
 import org.jabref.gui.DialogService;
-import org.jabref.gui.StateManager;
 import org.jabref.gui.icon.IconTheme;
 import org.jabref.gui.preview.PreviewPanel;
 import org.jabref.gui.theme.ThemeManager;
+import org.jabref.gui.util.OptionalObjectProperty;
 import org.jabref.gui.util.TaskExecutor;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.pdf.search.IndexingTaskManager;
+import org.jabref.logic.search.SearchQuery;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.preferences.PreferencesService;
@@ -17,26 +18,26 @@ public class PreviewTab extends EntryEditorTab implements OffersPreview {
     private final DialogService dialogService;
     private final BibDatabaseContext databaseContext;
     private final PreferencesService preferences;
-    private final StateManager stateManager;
     private final ThemeManager themeManager;
     private final IndexingTaskManager indexingTaskManager;
     private final TaskExecutor taskExecutor;
+    private final OptionalObjectProperty<SearchQuery> searchQueryProperty;
     private PreviewPanel previewPanel;
 
     public PreviewTab(BibDatabaseContext databaseContext,
                       DialogService dialogService,
                       PreferencesService preferences,
-                      StateManager stateManager,
                       ThemeManager themeManager,
                       IndexingTaskManager indexingTaskManager,
-                      TaskExecutor taskExecutor) {
+                      TaskExecutor taskExecutor,
+                      OptionalObjectProperty<SearchQuery> searchQueryProperty) {
         this.databaseContext = databaseContext;
         this.dialogService = dialogService;
         this.preferences = preferences;
-        this.stateManager = stateManager;
         this.themeManager = themeManager;
         this.indexingTaskManager = indexingTaskManager;
         this.taskExecutor = taskExecutor;
+        this.searchQueryProperty = searchQueryProperty;
 
         setGraphic(IconTheme.JabRefIcons.TOGGLE_ENTRY_PREVIEW.getGraphicNode());
         setText(Localization.lang("Preview"));
@@ -64,7 +65,7 @@ public class PreviewTab extends EntryEditorTab implements OffersPreview {
     @Override
     protected void bindToEntry(BibEntry entry) {
         if (previewPanel == null) {
-            previewPanel = new PreviewPanel(databaseContext, dialogService, preferences.getKeyBindingRepository(), preferences, stateManager, themeManager, indexingTaskManager, taskExecutor);
+            previewPanel = new PreviewPanel(databaseContext, dialogService, preferences.getKeyBindingRepository(), preferences, themeManager, indexingTaskManager, taskExecutor, searchQueryProperty);
             setContent(previewPanel);
         }